joystick inalambrico funcionando

+++ pocs/esp32-integracion-joystick/main/joystick_service.h
@@ -3,17 +3,15 @@
 #ifndef JOYSTICK_SERVICE_H
 #define JOYSTICK_SERVICE_H
 
-
-#include <stdio.h>
-
-
 #define JOYSTICK_READING_MAX 1
 #define JOYSTICK_READING_MIN -1
 #define JOYSTICK_READING_NEUTRAL 0
 
 
+#include <stdio.h>
+
 void joystick_initialize();
 
-void joystick_get_reading(float * read_x , float * read_y);
+void joystick_get_reading(int * read_x , int * read_y);
 
 #endif
